requirements.txt
gspread
google-auth
google-auth-oauthlib
google-api-python-client